requirements.txt
requests==2.31.0
python-dotenv==1.0.0
transformers==4.36.2
torch>=2.0.0
googletrans==4.0.0rc1
